+++ Cli/Command/Seed/SeedEmailBounceLog.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace TickTackk\Seeder\Cli\Command\Seed;
+
+use Symfony\Component\Console\Input\InputInterface;
+
+class SeedEmailBounceLog extends AbstractSeedCommand
+{
+    protected function getSeedName() : string
+    {
+        return 'email-bounce-log';
+    }
+
+    protected function getContentTypePlural(InputInterface $input = null) : string
+    {
+        return 'Email bounce logs';
+    }
+}